Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/xhr/formdata-blob-expected.txt


PASS formdata with blob 
PASS formdata with named blob 
PASS formdata.append() should throw if value is string and file name is given 
